requirements.txt
fastapi
uvicorn
python-dotenv
pymongo[srv]
sentence-transformers
transformers
torch
accelerate
huggingface_hub
certifi
